Add files via upload
Small Planetarium for Win32forth  , this is a learning tool to teach basic graphic animations in the console. You do not have to open any special windows to make graphics on Win32forth, this way you can concentrate your efforts learning Forth, not Windows !  This is the kind of programming we learned on the old DOS Forths, and in my opinion we had greater productivity concentrating to the real  purpose of the program and not wasting time with the environment for our program. It is also intended as a first steps for novices to advance in math , astronomy, and learning how to  draw  function  output to the screen.